+++ ios/Classes/core/PHAsset+PM_COMMON.h
@@ -27,9 +27,8 @@ NS_ASSUME_NONNULL_BEGIN
  @return The MIME type of this asset if available, otherwise `nil`.
  */
 - (nullable NSString*)mimeType;
-- (BOOL)isAdjust;
-- (PHAssetResource *)getAdjustResource;
-- (void)requestAdjustedData:(void (^)(NSData *_Nullable result))block;
+- (PHAssetResource *)getCurrentResource;
+- (void)requestCurrentResourceData:(void (^)(NSData *_Nullable result))block;
 - (PHAssetResource *)getLivePhotosResource;
 
 @end