+++ lib/src/managers/photo_manager.dart
@@ -68,7 +68,10 @@ class PhotoManager {
   /// See the documents from Apple:
   ///  * iOS 14: https://developer.apple.com/documentation/photokit/phphotolibrary/3616113-presentlimitedlibrarypickerfromv/
   ///  * iOS 15: https://developer.apple.com/documentation/photokit/phphotolibrary/3752108-presentlimitedlibrarypickerfromv/
-  static Future<void> presentLimited() => plugin.presentLimited();
+  static Future<void> presentLimited({
+    RequestType type = RequestType.all,
+  }) =>
+      plugin.presentLimited(type);
 
   /// Obtain albums/folders list with couple filter options.
   ///
@@ -124,6 +127,7 @@ class PhotoManager {
   /// call the [getAssetPathList] to start again.
   ///
   /// Make sure callers of this method have `await`ed properly.
+  /// The method does not supported on OpenHarmony.
   static Future<void> releaseCache() => plugin.releaseCache();
 
   /// {@macro photo_manager.NotifyManager.addChangeCallback}
@@ -168,6 +172,7 @@ class PhotoManager {
   static Future<String> systemVersion() => plugin.getSystemVersion();
 
   /// Clear all file caches.
+  /// The method does not supported on OpenHarmony.
   static Future<void> clearFileCache() => plugin.clearFileCache();
 
   /// Returns the count of assets.